% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gev.R
\name{gevreg.fit}
\alias{gevreg.fit}
\title{Maximum likelihood fit of the GEV regression model}
\usage{
gevreg.fit(X, y, xi_start = c(-0.2, -0.05, 0.05, 0.2))
}
\arguments{
\item{X}{Design matrix (including the intercept column).}

\item{y}{Response vector.}

\item{xi_start}{Vector of starting values for the shape parameter.}
}
\value{
A list with the estimates, the log-likelihood, the observed
  information (Hessian of the negative log-likelihood) and the convergence
  code.
}
\description{
Fits \eqn{Y_i \sim \mathrm{GEV}(\mu_i, \sigma, \xi)} with
\eqn{\mu_i = x_i^\top \beta} by maximum likelihood, using the analytic
gradient and several starting values for the shape parameter.
}
\details{
Centring the covariates is strongly recommended: the GEV likelihood is
poorly conditioned when covariates are far from the origin, and an
uncentred fit may fail to converge without any warning.
}
\examples{
data(itajai)
X <- cbind(1, itajai$pressure - mean(itajai$pressure))
gevreg.fit(X, itajai$wind)
}
