% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{plot.evbs_monitor}
\alias{plot.evbs_monitor}
\title{Plot an endpoint-identifiability control chart}
\usage{
\method{plot}{evbs_monitor}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{evbs_monitor}}.}

\item{...}{Further arguments passed to \code{plot}.}
}
\value{
Called for its side effect; invisibly returns \code{x}.
}
\description{
Plot an endpoint-identifiability control chart
}
