% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evbsreg_densities.R
\name{plot_evbs_alpha}
\alias{plot_evbs_alpha}
\title{Plot EVBS Densities for Varying alpha}
\usage{
plot_evbs_alpha()
}
\value{
A \code{ggplot} object.
}
\description{
Reproduces Figure 1(a) of the paper: EVBS densities for
\eqn{\alpha \in \{0.25, 0.5, 0.75, 1\}}, \eqn{\beta = 1},
\eqn{\gamma = 0.5}. Uses \pkg{ggplot2} with the Dark2 colour palette.
}
\examples{
\donttest{print(plot_evbs_alpha())}

}
\seealso{
\code{\link{plot_evbs_gama}}, \code{\link{generate_evbs_data}}.
}
