% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evbsreg_densities.R
\name{plot_evbs_gama}
\alias{plot_evbs_gama}
\title{Plot EVBS Densities for Varying gamma}
\usage{
plot_evbs_gama()
}
\value{
A \code{ggplot} object.
}
\description{
Reproduces Figure 1(b) of the paper: EVBS densities for
\eqn{\gamma \in \{-1.25, -0.75, 0.75, 1.25\}}, \eqn{\alpha = 0.5},
\eqn{\beta = 1.5}. Uses \pkg{ggplot2} with the Dark2 colour palette.
}
\examples{
\donttest{print(plot_evbs_gama())}

}
\seealso{
\code{\link{plot_evbs_alpha}}.
}
