% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evbsreg_fit.R
\name{print.evbsreg}
\alias{print.evbsreg}
\title{Print Method for EVBS Regression Fits}
\usage{
\method{print}{evbsreg}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"evbsreg"}.}

\item{digits}{Number of significant digits (default 4).}

\item{...}{Further arguments passed to \code{\link{print.default}}.}
}
\value{
The object \code{x}, invisibly. Called for the side effect of
  printing a coefficient table to the console.
}
\description{
Compactly prints the parameter estimates, standard errors, and Wald
tests of an object returned by \code{\link{evbsreg.fit}}.
}
\examples{
data(itajai)
fit <- evbsreg.fit(cbind(1, itajai$pressure), itajai$wind)
print(fit)

}
\seealso{
\code{\link{evbsreg.fit}}.
}
