% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evbsreg_residuals.R
\name{rcoxsnell}
\alias{rcoxsnell}
\title{Cox-Snell Residuals for the Log-EVBS Regression Model}
\usage{
rcoxsnell(X, t)
}
\arguments{
\item{X}{A numeric design matrix with an intercept column.}

\item{t}{A numeric vector of strictly positive responses.}
}
\value{
A numeric vector of length \code{length(t)} containing the
  Cox-Snell residuals.
}
\description{
Computes Cox-Snell residuals for a log-EVBS regression fit. Under a
correctly specified model these residuals form an approximately
standard exponential sample.
}
\examples{
data(itajai)
X <- cbind(1, itajai$pressure)
cs <- rcoxsnell(X, itajai$wind)
summary(cs)

}
\references{
Cox, D. R. and Snell, E. J. (1968). A general definition of residuals.
\emph{Journal of the Royal Statistical Society, Series B}, 30, 248--275.
}
\seealso{
\code{\link{rqrandomized}}, \code{\link{envelope_qq}}.
}
