% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evbsreg_residuals.R
\name{rqrandomized}
\alias{rqrandomized}
\title{Randomized Quantile Residuals for the Log-EVBS Regression Model}
\usage{
rqrandomized(X, t)
}
\arguments{
\item{X}{A numeric design matrix with an intercept column.}

\item{t}{A numeric vector of strictly positive responses.}
}
\value{
A numeric vector of length \code{length(t)} containing the
  randomized quantile residuals.
}
\description{
Computes randomized quantile residuals (Dunn and Smyth, 1996) for a
log-EVBS regression fit. Under a correctly specified model these
residuals are approximately standard normal, so departures from
normality indicate lack of fit.
}
\examples{
data(itajai)
X <- cbind(1, itajai$pressure)
r <- rqrandomized(X, itajai$wind)
shapiro.test(r)

}
\references{
Dunn, P. K. and Smyth, G. K. (1996). Randomized quantile residuals.
\emph{Journal of Computational and Graphical Statistics}, 5, 236--244.
}
\seealso{
\code{\link{rcoxsnell}}, \code{\link{envelope_qq}},
  \code{\link{evbsreg.fit}}.
}
